ticket:3667 for Tom Bendall: Ensure correct upwind direction for vertical theta transport.
git-svn-id: https://metomi/svn/lfric.xm/LFRic/trunk@40982 4bbb03d4-c1cc-11e1-9ca4-71d347b01581

+++ gungho/source/kernel/transport/mol/poly1d_vert_adv_kernel_mod.F90
@@ -161,7 +161,7 @@ subroutine poly1d_vert_adv_code( nlayers,              &
     ! reconstructions only.
     ! if wind > 0 -> upwind_offset = 1
     ! if wind < 0 -> upwind_offset = 0
-    upwind = int(0.5_r_tran*(1.0_r_tran + sign(1.0_r_tran,wind(map_w2v(1)+k))),i_def)
+    upwind = nint(0.5_r_tran*(1.0_r_tran + sign(1.0_r_tran,wind(map_w2v(1)+k))),i_def)
     upwind_offset = use_upwind*upwind
     stencil = stencil - upwind_offset
